requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
geographiclib==1.49
geopy==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pandas==0.20.3
pycodestyle==2.5.0
python-dotenv==0.10.2
requests==2.21.0
urllib3==1.24.3
Werkzeug==0.15.4
